refactor: move project papers view select components to components directory
